# schema/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/chats/": {
            "put": {
                "tags": [
                    "chats"
                ],
                "summary": "Edit Chat Data",
                "description": "Edit the info of the chat, add/remove users, change admins. Access: chat admins.",
                "operationId": "edit_chat_data_chats__put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EditChatDataRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EditChatDataResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "chats"
                ],
                "summary": "Create Chat",
                "description": "Create a new chat and add some users to it. The sender must be one of the invited users.\nThey will also be chat's admin.",
                "operationId": "create_chat_chats__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateChatRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateChatResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "chats"
                ],
                "summary": "Delete Chat",
                "description": "Delete the chat, including all its messages. Access: chat admins.",
                "operationId": "delete_chat_chats__delete",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeleteChatRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteChatResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/chats/{chat_id}": {
            "get": {
                "tags": [
                    "chats"
                ],
                "summary": "Get Chat Data",
                "description": "Get the data of the chat, including member users and admins. Access: chat members.",
                "operationId": "get_chat_data_chats__chat_id__get",
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Chat Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetChatDataResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/chatbot/": {
            "post": {
                "tags": [
                    "chatbot"
                ],
                "summary": "Chat Endpoint",
                "description": "Chat endpoint that handles user messages and returns responses from the chatbot.\nRequires an API token for authorization.\n\nParameters\n----------\nrequest : Request\n    The HTTP request object.\nchat_request : ChatRequest\n    The user message for the chatbot.\ntoken : str\n    The Bearer token for authentication.\n\nReturns\n-------\nChatResponse\n    The response from the chatbot.",
                "operationId": "chat_endpoint_chatbot__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ChatRequest": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "message"
                ],
                "title": "ChatRequest"
            },
            "ChatResponse": {
                "properties": {
                    "response": {
                        "type": "string",
                        "title": "Response"
                    }
                },
                "type": "object",
                "required": [
                    "response"
                ],
                "title": "ChatResponse"
            },
...
            "CreateChatRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags"
                    },
                    "image_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Id"
                    },
                    "users": {
                        "items": {
                            "type": "integer"
                        },
                        "type": "array",
                        "title": "Users"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "tags",
                    "users"
                ],
                "title": "CreateChatRequest",
                "description": "Create a new chat with users. Creator is the initial admin."
            },
            "CreateChatResponse": {
                "properties": {
                    "chat_id": {
                        "type": "integer",
                        "title": "Chat Id"
                    }
                },
                "type": "object",
                "required": [
                    "chat_id"
                ],
                "title": "CreateChatResponse"
            },
...
            "DeleteChatRequest": {
                "properties": {
                    "chat_id": {
                        "type": "integer",
                        "title": "Chat Id"
                    }
                },
                "type": "object",
                "required": [
                    "chat_id"
                ],
                "title": "DeleteChatRequest",
                "description": "Delete a chat, including all its messages."
            },
            "DeleteChatResponse": {
                "properties": {},
                "type": "object",
                "title": "DeleteChatResponse"
            },
...
            "EditChatDataRequest": {
                "properties": {
                    "chat_id": {
                        "type": "integer",
                        "title": "Chat Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags"
                    },
                    "image_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Id"
                    },
                    "users": {
                        "items": {
                            "type": "integer"
                        },
                        "type": "array",
                        "title": "Users"
                    },
                    "admin_users": {
                        "items": {
                            "type": "integer"
                        },
                        "type": "array",
                        "title": "Admin Users"
                    }
                },
                "type": "object",
                "required": [
                    "chat_id",
                    "name",
                    "description",
                    "tags",
                    "users",
                    "admin_users"
                ],
                "title": "EditChatDataRequest",
                "description": "Edit chat data. Can only be done by chat admin."
            },
            "EditChatDataResponse": {
                "properties": {},
                "type": "object",
                "title": "EditChatDataResponse"
            },
...
            "GetChatDataResponse": {
                "properties": {
                    "chat_id": {
                        "type": "integer",
                        "title": "Chat Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags"
                    },
                    "image_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Id"
                    },
                    "users_infos": {
                        "items": {
                            "$ref": "#/components/schemas/UserShortInfo"
                        },
                        "type": "array",
                        "title": "Users Infos"
                    },
                    "admin_users": {
                        "items": {
                            "type": "integer"
                        },
                        "type": "array",
                        "title": "Admin Users"
                    }
                },
                "type": "object",
                "required": [
                    "chat_id",
                    "name",
                    "tags",
                    "users_infos",
                    "admin_users"
                ],
                "title": "GetChatDataResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "UserShortInfo": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "image_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Id"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name"
                ],
                "title": "UserShortInfo"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "APIKeyHeader": {
                "type": "apiKey",
                "description": "Bearer token",
                "in": "header",
                "name": "Authorization"
            }
        }
    }
}